Permiso, Rol y Usuario, quedarian mappers

+++ src/main/java/com/example/finca_hexagonal/domain/models/Permiso.java
@@ -1,16 +1,16 @@
 package com.example.finca_hexagonal.domain.models;
 
 public class Permiso {
-    private int id;
+    private Long id;
     private String nombre;
 
-    public Permiso(int id, String nombre){
+    public Permiso(Long id, String nombre){
         this.id = id;
         this.nombre = nombre;
     }
     public Permiso(){}
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
@@ -18,7 +18,7 @@ public class Permiso {
         return nombre;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
